[states] mv State to new module, implement InGameState
